boot_qemu.py: wic issues with Zynq and SOM
SOM case: if we want to use USB for wic then the input file arg is
if=none for sd it is if=sd.
In python qemu we have added only sd now adding USB also.
For Zynq and others adding wic restricted only when --qemu-args mentioned.
Fixing that as if ext4 rootfs enabled and wic available then using
that for boot irrespective of --qemu-args.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>