plnx_vars:plnx_utils: add bootscr address in dtsi
In SDT flow bootscr address, in u-boot options node
missing. Adding the support for SDT flow as well.
U-boot checks for the bootscr in 0x20000000 location.
When we use upper ddr or HBM to load u-boot unable to
find the bootscr script. To handle this passing
bootscr address to u-boot through options node.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>